Version 02 - Incomplete
V02 (Dec 10, 2023): Second partial version of the app. The plotting with no subgrouping has been transformed to work in the Streamlit app, as well as the plot customizations. It is still pending to transform the option to create multiple subgroups, to save the plot + dataset, and to review the logging.